backend/requirements.txt
fastapi
uvicorn
gunicorn; sys_platform != 'win32'
motor
dnspython
redis
qdrant-client
python-dotenv
pydantic-settings
beanie
python-multipart
passlib[bcrypt]
python-jose[cryptography]
bcrypt
openai
email-validator
httpx
# Removed langchain to save memory on cloud free tiers
